requirements.txt
polygon-api-client>=1.12.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.10.0
lxml>=4.9.0
html5lib>=1.1
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0

# For gold predictor (economic data & ML)
fredapi>=0.5.0
scikit-learn>=1.3.0

# For gold tracker notifications (Windows only)
win10toast>=0.9; sys_platform == 'win32'

